Major reworking of field datatypes for the beam array data. Most fields decoded using 4 bytes will be converted to float64, whereas those decoded using 1 or 2 bytes will be converted to float32. Scale and offsets are applied in float64, and any required conversions to float32 or integers are applied later. Nulls have definitions for both float32 and float64 (where a float32 was defined, now a float64 is also defined. Also cater for different gsf versions to read different ping information.

+++ geo.go
@@ -76,8 +76,8 @@ func (ba *BeamArray) BeamsLonLat(lon, lat float64, heading float32, coef *GeoCoe
 	lat2 := make([]float64, n)
 
 	for i := 0; i < n; i++ {
-		acr_trck = float64(ba.AcrossTrack[i])
-		aln_trck = float64(ba.AlongTrack[i])
+		acr_trck = ba.AcrossTrack[i]
+		aln_trck = ba.AlongTrack[i]
 		lon2[i] = lon + delta_y/lon_sf*acr_trck + delta_x/lon_sf*aln_trck
 		lat2[i] = lat - delta_x/lat_sf*acr_trck + delta_y/lat_sf*aln_trck
 	}
